# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Nifty100 Financial Intelligence API","version":"1.0.0"},"paths":{"/api/v1/companies/":{"get":{"tags":["Companies"],"summary":"Get Companies","operationId":"get_companies_api_v1_companies__get","parameters":[{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"market_cap_category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Market Cap Category"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/pl":{"get":{"tags":["Companies"],"summary":"Get Profit Loss","operationId":"get_profit_loss_api_v1_companies__ticker__pl_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/bs":{"get":{"tags":["Companies"],"summary":"Get Balance Sheet","operationId":"get_balance_sheet_api_v1_companies__ticker__bs_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/cashflow":{"get":{"tags":["Companies"],"summary":"Get Cashflow","operationId":"get_cashflow_api_v1_companies__ticker__cashflow_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"From Year"}},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"To Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/ratios":{"get":{"tags":["Companies"],"summary":"Get Financial Ratios","operationId":"get_financial_ratios_api_v1_companies__ticker__ratios_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/tearsheet":{"get":{"tags":["Companies"],"summary":"Get Tearsheet","operationId":"get_tearsheet_api_v1_companies__ticker__tearsheet_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}":{"get":{"tags":["Companies"],"summary":"Get Company","operationId":"get_company_api_v1_companies__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/screener/":{"get":{"tags":["Screener"],"summary":"Screener","operationId":"screener_api_v1_screener__get","parameters":[{"name":"min_roe","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Roe"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors/":{"get":{"tags":["Sectors"],"summary":"Get Sectors","operationId":"get_sectors_api_v1_sectors__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/sectors/{sector}":{"get":{"tags":["Sectors"],"summary":"Get Sector Companies","operationId":"get_sector_companies_api_v1_sectors__sector__get","parameters":[{"name":"sector","in":"path","required":true,"schema":{"type":"string","title":"Sector"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/peers/":{"get":{"tags":["Peers"],"summary":"Peers Root","operationId":"peers_root_api_v1_peers__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/valuation/":{"get":{"tags":["Valuation"],"summary":"Valuation Root","operationId":"valuation_root_api_v1_valuation__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/portfolio/":{"get":{"tags":["Portfolio"],"summary":"Get Portfolios","operationId":"get_portfolios_api_v1_portfolio__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Portfolio"],"summary":"Create Portfolio","operationId":"create_portfolio_api_v1_portfolio__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/{portfolio_id}/holdings":{"post":{"tags":["Portfolio"],"summary":"Add Holding","operationId":"add_holding_api_v1_portfolio__portfolio_id__holdings_post","parameters":[{"name":"portfolio_id","in":"path","required":true,"schema":{"type":"integer","title":"Portfolio Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldingCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/{portfolio_id}":{"get":{"tags":["Portfolio"],"summary":"Get Portfolio","operationId":"get_portfolio_api_v1_portfolio__portfolio_id__get","parameters":[{"name":"portfolio_id","in":"path","required":true,"schema":{"type":"integer","title":"Portfolio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/{portfolio_id}/summary":{"get":{"tags":["Portfolio"],"summary":"Get Portfolio Summary","operationId":"get_portfolio_summary_api_v1_portfolio__portfolio_id__summary_get","parameters":[{"name":"portfolio_id","in":"path","required":true,"schema":{"type":"integer","title":"Portfolio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/{portfolio_id}/allocation":{"get":{"tags":["Portfolio"],"summary":"Get Portfolio Allocation","operationId":"get_portfolio_allocation_api_v1_portfolio__portfolio_id__allocation_get","parameters":[{"name":"portfolio_id","in":"path","required":true,"schema":{"type":"integer","title":"Portfolio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlist/":{"get":{"tags":["Watchlist"],"summary":"Get Watchlists","operationId":"get_watchlists_api_v1_watchlist__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Watchlist"],"summary":"Create Watchlist","operationId":"create_watchlist_api_v1_watchlist__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlist/{watchlist_id}/stocks":{"post":{"tags":["Watchlist"],"summary":"Add Stock","operationId":"add_stock_api_v1_watchlist__watchlist_id__stocks_post","parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlist/{watchlist_id}":{"get":{"tags":["Watchlist"],"summary":"Get Watchlist","operationId":"get_watchlist_api_v1_watchlist__watchlist_id__get","parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/watchlist/{watchlist_id}/stocks/{company_id}":{"delete":{"tags":["Watchlist"],"summary":"Remove Stock","operationId":"remove_stock_api_v1_watchlist__watchlist_id__stocks__company_id__delete","parameters":[{"name":"watchlist_id","in":"path","required":true,"schema":{"type":"integer","title":"Watchlist Id"}},{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/documents/":{"get":{"tags":["Documents"],"summary":"Documents Root","operationId":"documents_root_api_v1_documents__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/health/":{"get":{"tags":["Health"],"summary":"Health","description":"Health check endpoint.\n\nReturns:\n    dict: API status information.","operationId":"health_api_v1_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HoldingCreate":{"properties":{"company_id":{"type":"string","title":"Company Id"},"quantity":{"type":"number","title":"Quantity"},"average_price":{"type":"number","title":"Average Price"}},"type":"object","required":["company_id","quantity","average_price"],"title":"HoldingCreate"},"PortfolioCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"PortfolioCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WatchlistCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"WatchlistCreate"},"WatchlistItemCreate":{"properties":{"company_id":{"type":"string","title":"Company Id"}},"type":"object","required":["company_id"],"title":"WatchlistItemCreate"}}}}